// src/types/createWalletBody.ts
/**
 * Kriptonio Private API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: private
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import type { WalletType } from './walletType';

export interface CreateWalletBody {
  title: string;
  password: string;
  type: WalletType;
}
export namespace CreateWalletBody {}